Enable SkipPastNotifications
SkipPastNotifications is subtly broken somewhere in the code, but it
exists as a workaround to avoid storing state.

Ideally, the reminder should be stateful: notifications for events that
have not yet started should be stored so that they don't get emitted
again. The state is cleared once events start.

There should be multiple backends: one on-memory (default) and one
persistent (ideally using a very lightweight DB, e.g. a JSON file).